refactor/ TRANSLATE -> TRANSLATOR 로 이름 변경
