+++ src/api/translate.ts
@@ -1,4 +1,4 @@
-import TRANSLATE from '@/constants/TRANSLATE';
+import TRANSLATOR from '@/constants/TRANSLATOR';
 import getTranslateModel from '@/utils/getTranslateModel';
 import { Response } from '@/Interface/Response';
 import createResponse from '@/utils/createResponse';
@@ -14,7 +14,7 @@ export async function requestTranslatedText(text: string): Promise<Response> {
         const body = {
             message: text,
         };
-        const response = await model.post(TRANSLATE.TRANSLATE_URL, body);
+        const response = await model.post(TRANSLATOR.TRANSLATE_URL, body);
         const translatedText = response.data.body.text.translated;
 
         return createResponse(true, translatedText);
