+++ src/constants/TRANSLATOR/VIEW.ts
